created a flask blueprint for the endpoints to be served from one port
