+++ dynamic_webf/api/static/scripts/fvtailor.js
@@ -1,15 +1,15 @@
 $(document).ready(function(){
     // direct the user to the flair vibes tailor from the landing page whenever any of these buttons is clicked.
     $('#tryit_button').click(function() {
-        window.location = 'http://' + location.hostname + '/fvt';
+        window.location = 'http://' + location.hostname + ':5004/api/fvt';
     });
     $('#cta_button').click(function() {
-        window.location = 'http://' + location.hostname + '/fvt';
+        window.location = 'http://' + location.hostname + ':5004/api/fvt';
     });
 
 
     // code for the flair vibes tailor itself
-    const apiUrl = 'http://' + location.hostname + ':5006/submit_song';
+    const apiUrl = 'http://' + location.hostname + ':5004/api/submit_song';
     let query_data = {}
     $('form#submit_songs').submit(function(event) {
         event.preventDefault();
