+++ dynamic_webf/api/templates/fvtlandingpage.html
@@ -7,13 +7,13 @@
         <meta name="description" content=" Flair Vibes Tailor, a personalised music recommender that recommends similar songs based on what you submit as your favourite song.">
         <meta name="keywords" content="Flair Vibes Tailor, music, recommender, music recommendations, personalized music, music discovery, audio features">
         <title>FlairVibes Tailor</title>
-        <link rel="stylesheet" type="text/css" href="../fvtstatic/styles/common.css?{{ cache_id }}" />
-        <link rel="stylesheet" type="text/css" href="../fvtstatic/styles/header.css?{{ cache_id }}" />
-        <link rel="stylesheet" type="text/css" href="../fvtstatic/styles/footer.css?{{ cache_id }}" />
-        <LINK rel="icon" href="../fvtstatic/images/fvticon.png?{{ cache_id }}">
+        <link rel="stylesheet" type="text/css" href="../static/styles/common.css?{{ cache_id }}" />
+        <link rel="stylesheet" type="text/css" href="../static/styles/header.css?{{ cache_id }}" />
+        <link rel="stylesheet" type="text/css" href="../static/styles/footer.css?{{ cache_id }}" />
+        <LINK rel="icon" href="../static/images/fvticon.png?{{ cache_id }}">
     </head>
     <script src="https://code.jquery.com/jquery-3.2.1.min.js?{{ cache_id }}"></script>
-    <script src="../fvtstatic/scripts/fvtailor.js?{{ cache_id }}"></script>
+    <script src="../static/scripts/fvtailor.js?{{ cache_id }}"></script>
   <body>
     <header>
       <div class="logo">
